Mouse picking & weapon raycasting
Note to self: SDL lock cursor function caused the mouse picking to not working. I am still examing this problem and looking for a workaround this issue as I do want the mouse cursor to be locked in the middle of the screen. Additionally, I have created a debugger class to visually represent the ray to make it easier to debug it later on.